% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRG_plot_stat.R
\name{calculate_DRG_stats}
\alias{calculate_DRG_stats}
\title{Functions 2 for calculating statistics of the DRG data}
\usage{
calculate_DRG_stats(data, stat_type = c("mean", "median", "sd"))
}
\arguments{
\item{data}{a data frame}

\item{stat_type}{statistics over all of the DRG codes for average Medicare payments}
}
\value{
A statistics of \code{stat_type} by DRG code
}
\description{
Functions 2 for calculating statistics of the DRG data
}
\examples{
calculate_DRG_stats(drg_data, "mean")

}
